@@ -15,6 +15,7 @@
     "base64-js": "^1.5.1",
     "bootstrap": "^4.6.1",
     "compare-versions": "^5.0.0",
+    "crypto-browserify": "^3.12.0",
     "dayjs": "^1.10.7",
     "detect-browser": "^5.3.0",
     "fetch-retry": "^5.0.1",
@@ -23,15 +24,15 @@
     "libsodium-wrappers": "^0.7.9",
     "localforage": "^1.10.0",
     "page-lifecycle": "^0.1.2",
-    "react": "^17.0.2",
+    "query-string": "^7.1.1",
+    "react": "^18.2.0",
     "react-datepicker": "^4.0.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-js-pagination": "^3.0.3",
     "react-outside-click-handler": "^1.3.0",
     "react-redux": "^7.2.6",
     "react-redux-toastr": "^7.6.5",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
     "react-select": "^4.3.1",
     "react-with-direction": "^1.3.1",
     "reactstrap": "^8.9.0",
@@ -43,7 +44,15 @@
     "urijs": "^1.19.11",
     "uuid": "^9.0.0",
     "validator": "^13.7.0",
-    "xlsx": "^0.18.0"
+    "web-vitals": "^2.1.4",
+    "xlsx": "^0.18.5"
+  },
+  "devDependencies": {
+    "buffer": "^6.0.3",
+    "crypto": "npm:crypto-browserify",
+    "path": "npm:path-browserify",
+    "react-scripts": "5.0.1",
+    "stream": "npm:stream-browserify"
   },
   "scripts": {
     "dev": "react-scripts start -FAST_REFRESH=true",
