@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import 'react-redux-toastr/lib/css/react-redux-toastr.min.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.scss';
@@ -7,7 +7,8 @@ import App from './app';
 import './services/sentry';
 import './services/api';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const root = createRoot(document.getElementById('root'));
+root.render(<App tab="home" />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
